fix/#40: 퍼센테이지 채우기 에러 수정

+++ src/pages/main/MainPage.tsx
@@ -38,8 +38,8 @@ export const MainPage: React.FC = () => {
             )}
             collateralName={contract.copyright.name}
             presentValue={contract.copyright.ethPrice}
-            investmentProgressRate={parseInt(
-              contract.progress.replace(/[^0-9]/g, "")
+            investmentProgressRate={parseFloat(
+              contract.progress.replace(/[^0-9.]/g, "")
             )}
             onClick={() => navigate(`/investment/${contract.contractId}`)}
           />
